Fixed some errors in HASCO component target of sdlc: - PARAM_TYPE_INT is now 'int' - Parameter domain now correctly checks if interfaceParam definition is initialized.

+++ SEDE.core/servicedesc/src/main/java/de/upb/sede/util/ToHASCOComponentsTranslator.java
@@ -49,7 +49,7 @@ public class ToHASCOComponentsTranslator {
         PARAM_TYPE_BOOLEAN = "boolean",
         PARAM_TYPE_CATEGORY= "cat",
         PARAM_TYPE_DOUBLE = "double",
-        PARAM_TYPE_INT = "double"
+        PARAM_TYPE_INT = "int"
     ;
 
     private ToHASCOComponentsTranslator(ISDLLookupService sdlService,
